remove old SearchCommand remnants, move SearchCommand into qol commands
