--- server/src/main/kotlin/net/horizonsend/ion/server/features/client/whereisit/mod/FoundType.kt
@@ -1,11 +1,0 @@
-package net.horizonsend.ion.server.features.client.whereisit.mod
-
-/**
- * States how an item was found, used for different colours.
- * NOT_FOUND: The item was not found.
- * FOUND: The item was found directly.
- * FOUND_DEEP: The item was found in a nested inventory (shulker box, backpack).
- */
-enum class FoundType {
-	NOT_FOUND, FOUND, FOUND_DEEP
-}
