--- server/src/main/kotlin/net/horizonsend/ion/server/features/client/whereisit/mod/SearchResult.kt
@@ -1,5 +1,0 @@
-package net.horizonsend.ion.server.features.client.whereisit.mod
-
-import net.minecraft.network.chat.Component
-
-class SearchResult(val foundType: FoundType, val name: Component?)
